Add real E2E test for runWorkflow with FileChooser support
- Add run.test.ts testing full workflow including file attachment
- Add outlook-actions.test.ts for UI edge cases
- Update mock HTML with attach file menu for FileChooser event
- Colocate tests with source files in src/
- Move fixtures to src/__test__/fixtures/
- Remove redundant test files and unused scenarios
