+++ src/__test__/fixtures/signature.png
